Fix a big bug
Scraper was inserting strings like '08/17/2022 12:08' into a DATETIME
field.  sqlite3 cannot parse that as a date, but was willing to accept
the string into the column.  Consequently, I could not properly sort
by that column or compare it for inequality.

I have updated the scraper and added a script to fix old entries in
my database.